refactor: de-part trade screen cluster for wave 9 slice C (#4117)
Migrate the 20-part trade_screen library to explicit-import modules,
shrink the no-part grandfather by 21 entries, densify merged contract
and handler libraries for main-line budget compliance, and preserve
TRADE screen ID plus market/deal-book behavior.
