+++ app/lib/features/game/screens/trade/trade_screen.dart
@@ -1,118 +1,9 @@
 // Full-screen World Market Trade screen. SPEC/ui/trade-screen.md.
 //
-// **Scope (Refs #2993 E5a + E5b + E5c + E6):**
-// - E5a (already shipped) — route + screen ID + dark editorial-monocle
-//   chrome + observe-mode guard + two-tab body (Market + Deal Book) +
-//   read-only commodity table (last market price + previous-turn
-//   aggregate `Bids / Offers` volumes per commodity, sorted alphabetically
-//   by display name).
-// - E5b (already shipped) — interactive bid/offer/none direction selector
-//   and quantity stepper per row, wired to `currentOrdersProvider` so
-//   each change updates `Orders.tradeOrdersByPlayerId[player.id]` via the
-//   pure helpers `applyTradeOrderForPlayer` / `removeTradeOrderForPlayer`
-//   in `colonizethis_logic`. Mutual exclusion is structural: at most
-//   one staged `TradeOrder` per (player, commodityId) pair, so toggling
-//   from `Bid` to `Offer` (or vice versa) replaces the prior direction
-//   in place.
-// - E5c (already shipped) — cross-commodity cargo-remaining indicator +
-//   cap + warning. A persistent header strip above the commodity list
-//   renders `Cargo remaining: X` where
-//   `X = max(0, tradeCargoCapacity − totalBid)`. `tradeCargoCapacity` is
-//   `cargoHoldsForHomeFleet(game, player.id)` (falling back to
-//   `defaultCargoHoldsStub = 24` when the player has no home fleet).
-//   `totalBid` is the sum of staged `TradeOrderType.bid` quantities for
-//   the player; offers don't consume cargo (per #2988 §Cargo Constraint
-//   Model). When the cap is reached (`remainingCargo == 0 AND
-//   totalBid > 0`), a warning row is mounted below the indicator. Bid
-//   increments and `Bid` toggles are clamped so the cross-commodity bid
-//   total never exceeds `tradeCargoCapacity`.
-// - E6 (this slice) — Deal Book live two-panel ledger. The Deal Book tab
-//   body now renders the player's own previous-turn `FilledDeal`s and
-//   carry-forward orders, sourced from
-//   `Game.worldMarketState.lastTurnActivity[*].deals` (filtered by
-//   `buyerFactionId` / `sellerFactionId`) and
-//   `WorldMarketState.carryForward{Bids,Offers}ByFactionId[playerId]`.
-//   Two panels — `Your bids` (left, treasury spent) and `Your offers`
-//   (right, treasury received) — each with a Filled section and an
-//   Unfilled (carry-forward) section, plus a totals row that sums
-//   `quantity × pricePerUnit` over the filled deals only (carry-forwards
-//   have not cleared so they do not contribute to treasury totals). On
-//   viewports narrower than `dealBookTwoPanelMinWidth` (600 dp) the two
-//   panels stack vertically; otherwise they sit side-by-side. The Deal
-//   Book is purely read-only — observe-mode does not need a separate
-//   `IgnorePointer` wrap because the ledger has no interactive controls.
-//
-// Observe mode (the GP-observation `canMutateViaUi == false` case,
-// distinct from the "global observe / panels not defined" sentinel)
-// wraps the Market tab controls in `IgnorePointer` and visually dims
-// them so the table reads as read-only; the cargo indicator + warning
-// stay live (they read directly from `Game` + `currentOrdersProvider`).
-// The Deal Book ledger is rendered identically regardless of edit
-// permission — it always reflects the resolved-turn state.
-//
-// Deferred to follow-up slices:
-// - Priority dropdown (default priority 1 today; integration with the
-//   `kMaxTradePriority` API surface is tracked under #2989).
-//
-// The two-tab structure is the durable wireframe E6 builds inside, so
-// the contract this file ships matches what
-// `SPEC/ui/trade-screen.md` § Layout / wireframe records as the canonical
-// body for the screen.
+// De-parted wave-9 cluster (Refs #4117): explicit-import libraries replace the
+// former 20-part library. Public surface: [TradeScreen], [TradeScreenMarketKeys],
+// [TradeScreenDealBookKeys].
 
-import 'package:colonizethis_app_l10n/l10n/l10n.dart';
-import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_logic/colonizethis_logic.dart';
-import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-import '../../../../config/app_constants.dart';
-import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
-import '../../../../config/ui_screen_ids.dart';
-import '../../../../providers/games_provider.dart';
-import '../../../../providers/production_allocation_provider.dart';
-import '../../../../providers/treasury_summary_provider.dart';
-import '../../../../widgets/ct_choice_chip.dart';
-import '../../../../widgets/ct_game_feature_screen_shell.dart';
-import '../../../../widgets/ct_panel.dart';
-import '../../../../widgets/ct_section_label.dart';
-import '../../../../widgets/ct_spacing.dart';
-import '../../../../widgets/ct_tab_strip.dart';
-import '../../../../widgets/game_feature_screen_top_bar.dart';
-import '../../../../widgets/resource_icon.dart';
-import '../../../../widgets/strict_asset_icon.dart';
-import '../../widgets/production/commodity_ui_helpers.dart';
-import '../../widgets/shell/shell_player_context.dart';
-import '../../widgets/shell/shell_player_guarded_body.dart';
-import 'trade_section_handlers.dart';
-
-part 'trade_screen_contract_market.dart';
-part 'trade_screen_contract_market_rows.dart';
-part 'trade_screen_contract_deal_book.dart';
-part 'trade_screen_deal_book.dart';
-part 'trade_screen_deal_book_panel.dart';
-part 'trade_screen_deal_book_rows.dart';
-part 'trade_screen_market_row.dart';
-part 'trade_screen_market_row_controls.dart';
-part 'trade_screen_market_row_header.dart';
-part 'trade_screen_market_row_stepper.dart';
-part 'trade_screen_market_tab.dart';
-part 'trade_screen_market_tab_build.dart';
-part 'trade_screen_market_tab_build_sections.dart';
-part 'trade_screen_market_tab_cargo_header.dart';
-part 'trade_screen_market_tab_order_handlers.dart';
-part 'trade_screen_market_tab_order_handlers_direction.dart';
-part 'trade_screen_market_tab_order_handlers_quantity.dart';
-part 'trade_screen_market_tab_catalog.dart';
-part 'trade_screen_tabs_body.dart';
-part 'trade_screen_widget.dart';
-
-/// The Market tab body (`_MarketTabContent`, `_SectionedTradeableCommodities`)
-/// lives in the `trade_screen_market_tab.dart` part fragment, and its
-/// per-commodity row widgets (`_MarketCommodityRow`,
-/// `_MarketCommodityRowHeader`, `_MarketCommodityRowControls`,
-/// `_StepperButton`) live in the `trade_screen_market_row.dart` part
-/// fragment, keeping this host file under the repo code-review
-/// physical-line limit and the issue #3594 ≤700-line target
-/// (`SPEC/program/dart-file-non-comment-line-size.md`).
-
+export 'trade_screen_contract_deal_book.dart' show TradeScreenDealBookKeys;
+export 'trade_screen_contract_market.dart' show TradeScreenMarketKeys;
+export 'trade_screen_widget.dart' show TradeScreen;
