+++ app/lib/features/game/screens/trade/trade_screen_contract_deal_book.dart
@@ -1,7 +1,8 @@
-part of 'trade_screen.dart';
-
 /// Public Deal Book keys/literals for the trade screen.
 /// Tests and Deal Book UI parts use this type directly (Refs #4035 trade API collapse).
+
+import 'package:flutter/material.dart';
+
 abstract final class TradeScreenDealBookKeys {
   TradeScreenDealBookKeys._();
 
@@ -9,7 +10,7 @@ abstract final class TradeScreenDealBookKeys {
   /// tests asserting the Deal Book tab body is present in the tab
   /// strip's `IndexedStack` (visible when the Deal Book tab is selected
   /// after the user taps the Deal Book label). Refs #2993 E6 swapped
-  /// the placeholder for `_DealBookTabContent` — the key is intentionally
+  /// the placeholder for `DealBookTabContent` — the key is intentionally
   /// stable so existing tab-switch tests keep pinning the same body root.
   static const Key dealBookTabBodyKey =
       ValueKey<String>('tradeScreenDealBookTabBody');
