+++ app/lib/features/game/screens/trade/trade_screen_deal_book.dart
@@ -4,9 +4,7 @@
 // `SPEC/program/dart-file-non-comment-line-size.md`).
 //
 // All classes here are library-private (`_DealBook*`) and consumed
-// only by `_TradeScreenTabsBody` inside the parent library.
-
-part of 'trade_screen.dart';
+// only by `TradeScreenTabsBody` inside the parent library.
 
 /// Live Deal Book tab body (Refs #2993 E6). Renders the player's
 /// previous-turn buying and selling activity in a two-panel ledger
@@ -19,8 +17,16 @@ part of 'trade_screen.dart';
 /// stays overflow-safe (`SPEC/ui/mobile-adaptation.md` § 7). On wider
 /// viewports the bids panel sits left of the offers panel inside a
 /// `Row`.
-class _DealBookTabContent extends StatelessWidget {
-  const _DealBookTabContent({
+
+import 'package:flutter/material.dart';
+
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import 'trade_screen_contract_deal_book.dart';
+import 'trade_screen_deal_book_panel.dart';
+
+class DealBookTabContent extends StatelessWidget {
+  const DealBookTabContent({
     super.key,
     required this.game,
     required this.playerId,
@@ -31,7 +37,7 @@ class _DealBookTabContent extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final _DealBookViewData data = _DealBookViewData.build(
+    final DealBookViewData data = DealBookViewData.build(
       worldMarket: game.worldMarketState,
       playerId: playerId,
     );
@@ -52,8 +58,8 @@ class _DealBookTabContent extends StatelessWidget {
     );
   }
 
-  _DealBookPanel _buildBidsPanel(_DealBookViewData data) {
-    return _DealBookPanel(
+  DealBookPanel _buildBidsPanel(DealBookViewData data) {
+    return DealBookPanel(
       key: TradeScreenDealBookKeys.dealBookBidsPanelKey,
       panelTitle: TradeScreenDealBookKeys.dealBookBidsPanelTitle,
       side: TradeScreenDealBookKeys.dealBookSideBids,
@@ -67,8 +73,8 @@ class _DealBookTabContent extends StatelessWidget {
     );
   }
 
-  _DealBookPanel _buildOffersPanel(_DealBookViewData data) {
-    return _DealBookPanel(
+  DealBookPanel _buildOffersPanel(DealBookViewData data) {
+    return DealBookPanel(
       key: TradeScreenDealBookKeys.dealBookOffersPanelKey,
       panelTitle: TradeScreenDealBookKeys.dealBookOffersPanelTitle,
       side: TradeScreenDealBookKeys.dealBookSideOffers,
@@ -113,8 +119,8 @@ class _DealBookTabContent extends StatelessWidget {
 /// Deal Book view. Holds the four per-side row lists (filled / unfilled
 /// for bids and offers) and the two treasury totals. Pulled out so the
 /// rendering widget tree stays declarative and unit-testable.
-class _DealBookViewData {
-  const _DealBookViewData({
+class DealBookViewData {
+  const DealBookViewData({
     required this.filledBids,
     required this.filledOffers,
     required this.unfilledBids,
@@ -123,7 +129,7 @@ class _DealBookViewData {
     required this.totalReceived,
   });
 
-  factory _DealBookViewData.build({
+  factory DealBookViewData.build({
     required WorldMarketState worldMarket,
     required String playerId,
   }) {
@@ -143,7 +149,7 @@ class _DealBookViewData {
     for (final FilledDeal deal in offers) {
       received += deal.quantity * deal.pricePerUnit.floor();
     }
-    return _DealBookViewData(
+    return DealBookViewData(
       filledBids: List<FilledDeal>.unmodifiable(bids),
       filledOffers: List<FilledDeal>.unmodifiable(offers),
       unfilledBids:
