+++ app/lib/features/game/screens/trade/trade_screen_deal_book_panel.dart
@@ -1,7 +1,5 @@
 // Deal Book panel chrome + resolved styles for the World Market Trade screen.
 
-part of 'trade_screen.dart';
-
 /// Single ledger panel (one of `Your bids` / `Your offers`). Wraps the
 /// rows in a [CtPanel] so the dark editorial-monocle surface matches the
 /// sibling panels on this screen. Sections inside the panel:
@@ -15,8 +13,19 @@ part of 'trade_screen.dart';
 /// per-section headings collapse and a single empty-state line keyed
 /// [emptyKey] is rendered. The totals row remains mounted regardless so
 /// widget tests can pin the affordance.
-class _DealBookPanel extends StatelessWidget {
-  const _DealBookPanel({
+
+import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
+import 'package:flutter/material.dart';
+
+import 'package:colonizethis_logic/colonizethis_logic.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../../../../widgets/ct_panel.dart';
+import '../../../../widgets/ct_spacing.dart';
+import 'trade_screen_contract_deal_book.dart';
+
+class DealBookPanel extends StatelessWidget {
+  const DealBookPanel({
     super.key,
     required this.panelTitle,
     required this.side,
@@ -41,7 +50,7 @@ class _DealBookPanel extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final _DealBookPanelStyles styles = _DealBookPanelStyles.of(context);
+    final DealBookPanelStyles styles = DealBookPanelStyles.of(context);
     final bool panelEmpty = filledRows.isEmpty && unfilledRows.isEmpty;
     return CtPanel(
       padding: const EdgeInsets.all(CtSpacing.ml),
@@ -67,7 +76,7 @@ class _DealBookPanel extends StatelessWidget {
     );
   }
 
-  List<Widget> _buildSections(_DealBookPanelStyles styles) {
+  List<Widget> _buildSections(DealBookPanelStyles styles) {
     return <Widget>[
       Text(TradeScreenDealBookKeys.dealBookFilledHeading, style: styles.sectionHeading),
       const SizedBox(height: 4),
@@ -79,7 +88,7 @@ class _DealBookPanel extends StatelessWidget {
     ];
   }
 
-  List<Widget> _buildFilledRows(_DealBookPanelStyles styles) {
+  List<Widget> _buildFilledRows(DealBookPanelStyles styles) {
     if (filledRows.isEmpty) {
       return <Widget>[
         Text(TradeScreenDealBookKeys.dealBookFilledEmptyText, style: styles.muted),
@@ -89,7 +98,7 @@ class _DealBookPanel extends StatelessWidget {
       for (int i = 0; i < filledRows.length; i++)
         Padding(
           padding: EdgeInsets.only(top: i == 0 ? 0 : 2),
-          child: _DealBookFilledRow(
+          child: DealBookFilledRow(
             rowKey: TradeScreenDealBookKeys.dealBookFilledRowKey(side, i),
             deal: filledRows[i],
             rowStyle: styles.body,
@@ -99,7 +108,7 @@ class _DealBookPanel extends StatelessWidget {
     ];
   }
 
-  List<Widget> _buildUnfilledRows(_DealBookPanelStyles styles) {
+  List<Widget> _buildUnfilledRows(DealBookPanelStyles styles) {
     if (unfilledRows.isEmpty) {
       return <Widget>[
         Text(TradeScreenDealBookKeys.dealBookUnfilledEmptyText, style: styles.muted),
@@ -109,7 +118,7 @@ class _DealBookPanel extends StatelessWidget {
       for (int i = 0; i < unfilledRows.length; i++)
         Padding(
           padding: EdgeInsets.only(top: i == 0 ? 0 : 2),
-          child: _DealBookUnfilledRow(
+          child: DealBookUnfilledRow(
             rowKey: TradeScreenDealBookKeys.dealBookUnfilledRowKey(side, i),
             order: unfilledRows[i],
             rowStyle: styles.body,
@@ -120,9 +129,9 @@ class _DealBookPanel extends StatelessWidget {
 }
 
 /// Resolved per-panel text styles, isolated as a value object so the
-/// [_DealBookPanel] build path stays under the 60-line cap.
-class _DealBookPanelStyles {
-  const _DealBookPanelStyles({
+/// [DealBookPanel] build path stays under the 60-line cap.
+class DealBookPanelStyles {
+  const DealBookPanelStyles({
     required this.title,
     required this.sectionHeading,
     required this.body,
@@ -130,9 +139,9 @@ class _DealBookPanelStyles {
     required this.totals,
   });
 
-  factory _DealBookPanelStyles.of(BuildContext context) {
+  factory DealBookPanelStyles.of(BuildContext context) {
     final ThemeData theme = Theme.of(context);
-    return _DealBookPanelStyles(
+    return DealBookPanelStyles(
       title: (theme.textTheme.titleMedium ?? const TextStyle(fontSize: 16))
           .copyWith(color: EditorialMonoclePalette.accent),
       sectionHeading:
@@ -152,4 +161,95 @@ class _DealBookPanelStyles {
   final TextStyle body;
   final TextStyle muted;
   final TextStyle totals;
+}
+
+/// Single filled-deal row inside a Deal Book panel. Lays out
+/// `commodity — qty × price = notional` with optional FRR / FTP tags so
+/// the player can audit how the deal cleared per
+/// `SPEC/game/world-market.md` § Matching + § First Right of Refusal.
+
+
+
+
+class DealBookFilledRow extends StatelessWidget {
+  const DealBookFilledRow({
+    required this.rowKey,
+    required this.deal,
+    required this.rowStyle,
+    required this.tagStyle,
+  });
+
+  final Key rowKey;
+  final FilledDeal deal;
+  final TextStyle rowStyle;
+  final TextStyle tagStyle;
+
+  // ignore: avoid_hardcoded_strings_in_widgets
+  static const String _frrTag = 'FRR';
+  // ignore: avoid_hardcoded_strings_in_widgets
+  static const String _ftpTag = 'FTP';
+
+  @override
+  Widget build(BuildContext context) {
+    final int unitPrice = deal.pricePerUnit.floor();
+    final int notional = deal.quantity * unitPrice;
+    final String priceText = TradeScreenDealBookKeys.formatFilledDealUnitPrice(
+      deal.pricePerUnit,
+    );
+    final List<String> tags = <String>[
+      if (deal.isFirstRightOfRefusalMatch) _frrTag,
+      if (deal.isFtpMatch) _ftpTag,
+    ];
+    return Row(
+      crossAxisAlignment: CrossAxisAlignment.baseline,
+      textBaseline: TextBaseline.alphabetic,
+      key: rowKey,
+      children: <Widget>[
+        Expanded(
+          child: Text(
+            // ignore: avoid_hardcoded_strings_in_widgets
+            '${deal.commodityId} — qty ${deal.quantity} × $priceText '
+            '= $notional',
+            style: rowStyle,
+            overflow: TextOverflow.ellipsis,
+          ),
+        ),
+        if (tags.isNotEmpty) ...<Widget>[
+          const SizedBox(width: 6),
+          Text(
+            // ignore: avoid_hardcoded_strings_in_widgets
+            tags.join(' '),
+            style: tagStyle,
+          ),
+        ],
+      ],
+    );
+  }
+}
+
+/// Single carry-forward order row inside a Deal Book panel. The order
+/// has not cleared yet so there is no per-unit price or notional —
+/// `commodity — qty N (priority P)` is the canonical readout.
+class DealBookUnfilledRow extends StatelessWidget {
+  const DealBookUnfilledRow({
+    required this.rowKey,
+    required this.order,
+    required this.rowStyle,
+  });
+
+  final Key rowKey;
+  final TradeOrder order;
+  final TextStyle rowStyle;
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(
+      // ignore: avoid_hardcoded_strings_in_widgets
+      '${order.commodityId} — qty ${order.quantity} '
+      '(priority ${order.priority})',
+      key: rowKey,
+      style: rowStyle,
+      overflow: TextOverflow.ellipsis,
+    );
+  }
 }
