+++ app/lib/features/game/screens/trade/trade_screen_market_row_header.dart
@@ -2,13 +2,21 @@
 // Split from `trade_screen_market_row.dart` to keep each trade-screen
 // part under the repo file-size target (Refs #3878).
 
-part of 'trade_screen.dart';
-
 /// Resource icon, commodity name, sellable `(N)` headroom readout, then a
 /// fixed-width trailing column that right-aligns the coin icon and price so
 /// every row shares the same price-digit column edge.
-class _MarketCommodityRowHeader extends StatelessWidget {
-  const _MarketCommodityRowHeader({
+
+import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
+import 'package:flutter/material.dart';
+
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../../../../widgets/resource_icon.dart';
+import '../../../../widgets/strict_asset_icon.dart';
+import 'trade_screen_contract_market.dart';
+
+class MarketCommodityRowHeader extends StatelessWidget {
+  const MarketCommodityRowHeader({
     required this.commodityId,
     required this.commodityDisplayName,
     required this.priceText,
