+++ app/lib/features/game/screens/trade/trade_screen_market_row_stepper.dart
@@ -2,13 +2,15 @@
 // Split from `trade_screen_market_row.dart` to keep each trade-screen
 // part under the repo file-size target (Refs #3878).
 
-part of 'trade_screen.dart';
-
-/// Compact stepper button used by [_MarketCommodityRowControls]. Uses
+/// Compact stepper button used by [MarketCommodityRowControls]. Uses
 /// an `InkWell` so the chrome stays in the editorial-monocle palette
 /// (no Material elevated buttons or accent splash colours).
-class _StepperButton extends StatelessWidget {
-  const _StepperButton({
+
+import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
+import 'package:flutter/material.dart';
+
+class StepperButton extends StatelessWidget {
+  const StepperButton({
     required this.buttonKey,
     required this.glyph,
     required this.semanticLabel,
