+++ app/lib/features/game/screens/trade/trade_screen_market_tab.dart
@@ -10,13 +10,11 @@
 // `trade_screen_market_tab_catalog.dart`, and
 // `trade_screen_market_tab_build.dart` (Refs #3878).
 //
-// All classes here are library-private (`_MarketTabContent`,
-// `_SectionedTradeableCommodities`) and consumed only by
-// `_TradeScreenTabsBody` inside the parent library, so they keep using
-// `TradeScreen` static constants and the sibling `_MarketCommodityRow`
+// All classes here are library-private (`MarketTabContent`,
+// `SectionedTradeableCommodities`) and consumed only by
+// `TradeScreenTabsBody` inside the parent library, so they keep using
+// `TradeScreen` static constants and the sibling `MarketCommodityRow`
 // part fragment without further plumbing.
-
-part of 'trade_screen.dart';
 
 /// Interactive commodity table for the Market tab (Refs #2993 E5a + E5b).
 ///
@@ -43,8 +41,20 @@ part of 'trade_screen.dart';
 /// deterministic for widget tests and Widgetbook stories. The list is
 /// scrollable (the cargo indicator header from Refs #2993 E5c lands
 /// above the list when its plumbing arrives — Refs #2988 §UI Design).
-class _MarketTabContent extends ConsumerWidget {
-  const _MarketTabContent({
+
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import 'package:colonizethis_logic/colonizethis_logic.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../../../../providers/games_provider.dart';
+import '../../../../providers/production_allocation_provider.dart';
+import '../../../../providers/treasury_summary_provider.dart';
+import 'trade_screen_market_tab_build.dart';
+
+class MarketTabContent extends ConsumerWidget {
+  const MarketTabContent({
     super.key,
     required this.game,
     required this.playerId,
