+++ app/lib/features/game/screens/trade/trade_screen_market_tab_cargo_header.dart
@@ -2,16 +2,19 @@
 // Split from `trade_screen_market_tab.dart` to keep the host `build`
 // body under the repo file-size target (Refs #3878).
 
-part of 'trade_screen.dart';
-
 /// Persistent header above the Market commodity list (Refs #2993 E5c).
 ///
 /// Renders `Cargo remaining: X` and, when the cross-commodity bid cap is
 /// saturated, the cargo-limit warning line. Shared by the editable and
 /// observe-mode bodies so telemetry stays live even when row controls are
 /// dimmed.
-class _MarketTabCargoHeader extends StatelessWidget {
-  const _MarketTabCargoHeader({
+
+import 'package:flutter/material.dart';
+
+import 'trade_screen_contract_market.dart';
+
+class MarketTabCargoHeader extends StatelessWidget {
+  const MarketTabCargoHeader({
     required this.clampedRemaining,
     required this.warningVisible,
     required this.cargoIndicatorStyle,
